[#59] fix: 言語を変更すると、selectのタイトルが前の言語のままになる問題の修正
